pyproject.toml
[tool.poetry]
name = "laugh-score-analysis"
version = "0.1.0"
description = "笑い声分析ツール"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
google-genai = "^0.2.2"
ipython = "^8.22.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
